interface

+++ method.go
@@ -1,13 +1,13 @@
 package main
 
 type rect struct {
-	width, hieght int
+	width, hieght float64
 }
 
-func (r *rect) area() int {
+func (r *rect) area() float64 {
 	return r.width * r.hieght
 }
 
-func (r *rect) perim() int {
+func (r *rect) perim() float64 {
 	return 2*r.width + 2*r.hieght
 }
